feat: small review adjustments and consolidated deploy script
